Make sure dcache, MMU and IRQ are disabled at boot
Upon reset, the CPU is in a known state whereby data cache and MMU _ARE_
disabled. But when branching to address 0x8000, the start of our kernel,
we have no idea of what has happened just before. Who had the CPU before
us? Was it a previously running kernel running kexec to launch us? Was
it a U-boot bootloader?

As we can't just assume that the program calling us did the correct job,
we don't make any assumptions on the CPU state.